requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distro==1.8.0
dj-rest-auth==4.0.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
gTTS==2.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
inflection==0.5.1
install==1.3.5
openai==1.5.0
packaging==23.1
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
